routine and activity updates
Ahora se muestran las pantallas de rutinas y actividades como en el prototipo (a nivel backend)
